Slice 12: Split useAuth into its own file and fix AuthContext memoization
- Move AuthContext object to authContextValue.js (private internal module).
- Move useAuth hook to its own useAuth.js file so AuthContext.jsx exports
  only AuthProvider (kills react-refresh/only-export-components).
- Wrap login/devLogin/logout/hasRole in useCallback and add them to the
  useMemo deps so the context value is correctly memoized.
- Update all 11 useAuth import sites to point at ../auth/useAuth.
- Add file-header docblocks and JSDoc to all three auth files.
