+++ src/admin-app/src/features/groups/GroupsPage.jsx
@@ -24,7 +24,7 @@ import {
 
 import ConfirmDialog from "../../components/shared/ConfirmDialog";
 import { getGroups, getUsers, getChallenges, deleteGroup, logActivity } from "../../data/api";
-import { useAuth } from "../auth/AuthContext";
+import { useAuth } from "../auth/useAuth";
 
 export default function GroupsPage() {
   const [groups, setGroups] = useState([]);
