+++ src/admin-app/src/features/presets/PresetsPage.jsx
@@ -26,7 +26,7 @@ import {
 
 import ConfirmDialog from "../../components/shared/ConfirmDialog";
 import { getPresets, deletePreset } from "../../data/api";
-import { useAuth } from "../auth/AuthContext";
+import { useAuth } from "../auth/useAuth";
 
 export default function PresetsPage() {
   const [presets, setPresets] = useState([]);
